clickControl: validate key/value to updateClickConfig
updateClickConfig takes basically anything you want, send it a list,
dict, a fish.  Click is fine with any input.  The issue is that when
a user wants to change the latency of a click element, the user will
not know if the inputs from the user that click accepted, are valid
or work.

This commit is a halfway there commit, it may not be possible to
validate that click made the correct configuration change, but it is
easy to check if the click element or key they used were valid.

Hopefully in future commits, an additional check will be made on the
value, such that latency changes have the suffix .

Lincoln